# Optimize foldl/foreach for zip(arrays...), CartesianIndices, etc.