--- a/base/abstractarray.jl
+++ b/base/abstractarray.jl
@@ -1925,8 +1925,7 @@ julia> foreach(x -> println(x^2), a)
 ```
 """
 foreach(f) = (f(); nothing)
-foreach(f, itr) = (for x in itr; f(x); end; nothing)
-foreach(f, itrs...) = (for z in zip(itrs...); f(z...); end; nothing)
+# actual implementations are in reduce.jl
 
 ## map over arrays ##